requirements.txt
Django==4.1.5
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-filter==22.1
rest-framework-roles
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
requests
drf-yasg==1.21.5
pandas==1.5.3
whitenoise
openai
xhtml2pdf
django_smtp_ssl
rest_framework_roles
openpyxl==3.1.0
gunicorn
boto3
selenium
apscheduler